+++ trunk/ terasoft-12/Mechanect/Mechanect/Mechanect/Classes/Prey.cs
@@ -23,8 +23,8 @@ namespace Mechanect.Classes
             }
         }
 
-        int length;
-        public int Length
+        float length;
+        public float Length
         {
             get
             {
@@ -35,8 +35,8 @@ namespace Mechanect.Classes
                 length = value;
             }
         }
-        int width;
-        public int Width
+        float width;
+        public float Width
         {
             get
             {
@@ -48,7 +48,7 @@ namespace Mechanect.Classes
             }
         }
 
-        public Prey(Vector2 location, int width, int length)
+        public Prey(Vector2 location, float width, float length)
         {
             this.location = location;
             this.length = length;
@@ -73,7 +73,7 @@ namespace Mechanect.Classes
         /// <para>AUTHOR: Tamer Nabil </para>
         /// </remarks>
         /// <returns></returns>
-        public int getHeight()
+        public float getHeight()
         {
             return length;
         }
@@ -84,7 +84,7 @@ namespace Mechanect.Classes
         /// <para>AUTHOR: Tamer Nabil </para>
         /// </remarks>
         /// <returns></returns>
-        public int getWidth()
+        public float getWidth()
         {
             return width;
         }
